@@ -6,27 +6,34 @@
       "ByJG\\AnyDataset\\Db\\": "src/"
     }
   },
+    "autoload-dev": {
+      "psr-4": {
+        "Test\\": "tests/",
+        "TestDb\\": "testsdb/"
+      }
+    },
   "prefer-stable": true,
   "minimum-stability": "dev",
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1 <8.4",
     "ext-pdo": "*",
-    "byjg/anydataset-array": "4.9.*",
-    "psr/cache": "^1.0",
-    "psr/log": "^1.0",
-    "byjg/uri": "4.9.*",
+    "psr/cache": "^1.0|^2.0|^3.0",
+    "psr/log": "^1.1|^2.0|^3.0",
+    "byjg/uri": "^5.0",
+    "byjg/anydataset-array": "^5.0",
     "ext-json": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.7.*|7.4.*|^9.5",
-    "byjg/cache-engine": "4.9.*"
+    "phpunit/phpunit": "^9.6",
+    "vimeo/psalm": "^5.9",
+    "byjg/cache-engine": "^5.0"
   },
   "suggest": {
     "ext-curl": "*",
     "ext-pdo_mysql": "*",
     "ext-pdo_pgsql": "*",
     "ext-oci8": "*",
-    "byjg/cache-engine": "4.9.*"
+    "byjg/cache-engine": "^5.0"
   },
   "provide": {
     "byjg/anydataset-implementation": "1.0"
