@@ -13,18 +13,18 @@ class DbIterator extends GenericIterator
 
     const RECORD_BUFFER = 50;
 
-    private $rowBuffer;
-    private $currentRow = 0;
+    private array $rowBuffer;
+    private int $currentRow = 0;
 
     /**
-     * @var PDOStatement
+     * @var PDOStatement|null
      */
-    private $statement;
+    private ?PDOStatement $statement;
 
     /**
      * @param PDOStatement $recordset
      */
-    public function __construct($recordset)
+    public function __construct(PDOStatement $recordset)
     {
         $this->statement = $recordset;
         $this->rowBuffer = array();
@@ -34,7 +34,7 @@ class DbIterator extends GenericIterator
      * @return int
      */
     #[\ReturnTypeWillChange]
-    public function count()
+    public function count(): int
     {
         return $this->statement->rowCount();
     }
@@ -43,7 +43,7 @@ class DbIterator extends GenericIterator
      * @return bool
      * @throws InvalidArgumentException
      */
-    public function hasNext()
+    public function hasNext(): bool
     {
         if (count($this->rowBuffer) >= DbIterator::RECORD_BUFFER) {
             return true;
@@ -72,10 +72,10 @@ class DbIterator extends GenericIterator
     }
 
     /**
-     * @return Row
+     * @return Row|null
      * @throws InvalidArgumentException
      */
-    public function moveNext()
+    public function moveNext(): ?Row
     {
         if (!$this->hasNext()) {
             return null;
@@ -86,7 +86,7 @@ class DbIterator extends GenericIterator
         return $singleRow;
     }
 
-    public function key()
+    public function key(): int
     {
         return $this->currentRow;
     }
