@@ -4,12 +4,14 @@ namespace ByJG\AnyDataset\Db;
 
 use ByJG\AnyDataset\Core\Exception\DatabaseException;
 use ByJG\AnyDataset\Core\Exception\NotImplementedException;
+use ByJG\AnyDataset\Core\GenericIterator;
 use ByJG\AnyDataset\Db\Exception\DbDriverNotConnected;
 use ByJG\AnyDataset\Db\Helpers\SqlBind;
 use ByJG\AnyDataset\Db\Helpers\SqlHelper;
 use ByJG\AnyDataset\Db\Traits\DbCacheTrait;
 use ByJG\AnyDataset\Db\Traits\TransactionTrait;
 use ByJG\Util\Uri;
+use DateInterval;
 use Exception;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
@@ -22,9 +24,9 @@ class DbOci8Driver implements DbDriverInterface
 
     private LoggerInterface $logger;
 
-    private DbFunctionsInterface $dbHelper;
+    private ?DbFunctionsInterface $dbHelper = null;
 
-    public static function schema()
+    public static function schema(): array
     {
         return ['oci8'];
     }
@@ -34,11 +36,13 @@ class DbOci8Driver implements DbDriverInterface
      *
      * @var Uri
      */
-    protected $connectionUri;
+    protected Uri $connectionUri;
 
-    /** Used for OCI8 connections * */
-    protected $conn;
-    protected $ociAutoCommit;
+    /**
+     * @var resource|false
+     */
+    protected mixed $conn;
+    protected int $ociAutoCommit;
 
     /**
      * Ex.
@@ -52,6 +56,7 @@ class DbOci8Driver implements DbDriverInterface
     {
         $this->logger = new NullLogger();
         $this->connectionUri = $connectionString;
+        /** @psalm-suppress UndefinedConstant */
         $this->ociAutoCommit = OCI_COMMIT_ON_SUCCESS;
         $this->reconnect();
     }
@@ -61,7 +66,7 @@ class DbOci8Driver implements DbDriverInterface
      * @param Uri $connUri
      * @return string
      */
-    public static function getTnsString(Uri $connUri)
+    public static function getTnsString(Uri $connUri): string
     {
         $protocol = $connUri->getQueryPart("protocol");
         $protocol = ($protocol == "") ? 'TCP' : $protocol;
@@ -87,12 +92,13 @@ class DbOci8Driver implements DbDriverInterface
     }
 
     /**
-     * @param $sql
-     * @param null $array
+     * @param string $sql
+     * @param array|null $array
      * @return resource
      * @throws DatabaseException
+     * @throws DbDriverNotConnected
      */
-    protected function getOci8Cursor($sql, $array = null)
+    protected function getOci8Cursor(string $sql, array $array = null)
     {
         if (is_null($this->conn)) {
             throw new DbDriverNotConnected('Instance not connected');
@@ -129,12 +135,15 @@ class DbOci8Driver implements DbDriverInterface
     }
 
     /**
-     * @param $sql
-     * @param null $params
-     * @return Oci8Iterator
+     * @param string $sql
+     * @param array|null $params
+     * @param CacheInterface|null $cache
+     * @param int|DateInterval $ttl
+     * @return GenericIterator
      * @throws DatabaseException
+     * @throws DbDriverNotConnected
      */
-    public function getIterator($sql, $params = null, CacheInterface $cache = null, $ttl = 60)
+    public function getIterator(string $sql, ?array $params = null, ?CacheInterface $cache = null, DateInterval|int $ttl = 60): GenericIterator
     {
         return $this->getIteratorUsingCache($sql, $params, $cache, $ttl, function ($sql, $params) {
             $cur = $this->getOci8Cursor($sql, $params);
@@ -143,20 +152,22 @@ class DbOci8Driver implements DbDriverInterface
     }
 
     /**
-     * @param $sql
-     * @param null $array
-     * @return null
+     * @param string $sql
+     * @param array|null $array
+     * @return mixed
      * @throws DatabaseException
+     * @throws DbDriverNotConnected
      */
-    public function getScalar($sql, $array = null)
+    public function getScalar(string $sql, ?array $array = null): mixed
     {
         $cur = $this->getOci8Cursor($sql, $array);
 
+        /** @psalm-suppress UndefinedConstant */
         $row = oci_fetch_array($cur, OCI_RETURN_NULLS);
         if ($row) {
             $scalar = $row[0];
         } else {
-            $scalar = null;
+            $scalar = false;
         }
 
         oci_free_cursor($cur);
@@ -165,11 +176,12 @@ class DbOci8Driver implements DbDriverInterface
     }
 
     /**
-     * @param $tablename
+     * @param string $tablename
      * @return array
      * @throws DatabaseException
+     * @throws DbDriverNotConnected
      */
-    public function getAllFields($tablename)
+    public function getAllFields(string $tablename): array
     {
         $cur = $this->getOci8Cursor(SqlHelper::createSafeSQL("select * from :table", array(':table' => $tablename)));
 
@@ -185,19 +197,22 @@ class DbOci8Driver implements DbDriverInterface
         return $fields;
     }
 
-    protected function transactionHandler($action, $isolLevelCommand = "")
+    protected function transactionHandler(TransactionStageEnum $action, string $isoLevelCommand = ""): void
     {
         switch ($action) {
-            case 'begin':
+            case TransactionStageEnum::begin:
+                /** @psalm-suppress UndefinedConstant */
                 $this->ociAutoCommit = OCI_NO_AUTO_COMMIT;
-                $this->execute($isolLevelCommand);
+                $this->execute($isoLevelCommand);
                 break;
 
-            case 'commit':
+            case TransactionStageEnum::commit:
+                /** @psalm-suppress UndefinedConstant */
                 if ($this->ociAutoCommit == OCI_COMMIT_ON_SUCCESS) {
                     throw new DatabaseException('No transaction for commit');
                 }
 
+                /** @psalm-suppress UndefinedConstant */
                 $this->ociAutoCommit = OCI_COMMIT_ON_SUCCESS;
 
                 $result = oci_commit($this->conn);
@@ -207,11 +222,13 @@ class DbOci8Driver implements DbDriverInterface
                 }
                 break;
 
-            case 'rollback':
+            case TransactionStageEnum::rollback:
+                /** @psalm-suppress UndefinedConstant */
                 if ($this->ociAutoCommit == OCI_COMMIT_ON_SUCCESS) {
                     throw new DatabaseException('No transaction for rollback');
                 }
 
+                /** @psalm-suppress UndefinedConstant */
                 $this->ociAutoCommit = OCI_COMMIT_ON_SUCCESS;
 
                 oci_rollback($this->conn);
@@ -220,12 +237,12 @@ class DbOci8Driver implements DbDriverInterface
     }
     
     /**
-     * @param $sql
-     * @param null $array
+     * @param string $sql
+     * @param array|null $array
      * @return bool
      * @throws DatabaseException
      */
-    public function execute($sql, $array = null)
+    public function execute(string $sql, ?array $array = null): bool
     {
         $cur = $this->getOci8Cursor($sql, $array);
         oci_free_cursor($cur);
@@ -234,47 +251,46 @@ class DbOci8Driver implements DbDriverInterface
 
     /**
      *
-     * @return resource
+     * @return resource|false
      */
-    public function getDbConnection()
+    public function getDbConnection(): mixed
     {
         return $this->conn;
     }
 
     /**
-     * @param $name
+     * @param string $name
      * @throws NotImplementedException
      */
-    public function getAttribute($name)
+    public function getAttribute(string $name): mixed
     {
         throw new NotImplementedException('Method not implemented for OCI Driver');
     }
 
     /**
-     * @param $name
-     * @param $value
+     * @param string $name
+     * @param mixed $value
      * @throws NotImplementedException
      */
-    public function setAttribute($name, $value)
+    public function setAttribute(string $name, mixed $value): void
     {
         throw new NotImplementedException('Method not implemented for OCI Driver');
     }
 
     /**
-     * @param $sql
-     * @param null $array
+     * @param string $sql
+     * @param array|null $array
      * @throws NotImplementedException
      */
-    public function executeAndGetId($sql, $array = null)
+    public function executeAndGetId(string $sql, ?array $array = null): mixed
     {
         return $this->getDbHelper()->executeAndGetInsertedId($this, $sql, $array);
     }
 
     /**
-     * @return \ByJG\AnyDataset\Db\DbFunctionsInterface|void
-     * @throws NotImplementedException
+     * @return DbFunctionsInterface
      */
-    public function getDbHelper()
+    public function getDbHelper(): DbFunctionsInterface
     {
         if (empty($this->dbHelper)) {
             $this->dbHelper = Factory::getDbFunctions($this->getUri());
@@ -285,7 +301,7 @@ class DbOci8Driver implements DbDriverInterface
     /**
      * @return Uri
      */
-    public function getUri()
+    public function getUri(): Uri
     {
         return $this->connectionUri;
     }
@@ -293,21 +309,21 @@ class DbOci8Driver implements DbDriverInterface
     /**
      * @throws NotImplementedException
      */
-    public function isSupportMultRowset()
+    public function isSupportMultiRowset(): bool
     {
         return false;
     }
 
     /**
-     * @param $multipleRowSet
+     * @param bool $multipleRowSet
      * @throws NotImplementedException
      */
-    public function setSupportMultRowset($multipleRowSet)
+    public function setSupportMultiRowset(bool $multipleRowSet): void
     {
         throw new NotImplementedException('Method not implemented for OCI Driver');
     }
 
-    public function reconnect($force = false)
+    public function reconnect(bool $force = false): bool
     {
         if ($this->isConnected() && !$force) {
             return false;
@@ -321,18 +337,14 @@ class DbOci8Driver implements DbDriverInterface
         $codePage = ($codePage == "") ? 'UTF8' : $codePage;
         $tns = DbOci8Driver::getTnsString($this->connectionUri);
         $connType = $this->connectionUri->getQueryPart("conntype");
-        switch ($connType) {
-            case "persistent":
-                $connectMethod = "oci_pconnect";
-                break;
-            case "new":
-                $connectMethod = "oci_new_connect";
-                break;
-            default:
-                $connectMethod = "oci_connect";
-                break;
-        }
 
+        $connectMethod = match ($connType) {
+            "persistent" => "oci_pconnect",
+            "new" => "oci_new_connect",
+            default => "oci_connect",
+        };
+
+        /** @psalm-suppress UndefinedConstant */
         $this->conn = $connectMethod(
             $this->connectionUri->getUsername(),
             $this->connectionUri->getPassword(),
@@ -349,13 +361,13 @@ class DbOci8Driver implements DbDriverInterface
         return true;
     }
 
-    public function disconnect()
+    public function disconnect(): void
     {
         $this->clearCache();
         $this->conn = null;
     }
 
-    public function isConnected($softCheck = false, $throwError = false)
+    public function isConnected(bool $softCheck = false, bool $throwError = false): bool
     {
         if (empty($this->conn)) {
             if ($throwError) {
@@ -380,12 +392,12 @@ class DbOci8Driver implements DbDriverInterface
         return true;
     }
 
-    public function enableLogger(LoggerInterface $logger)
+    public function enableLogger(LoggerInterface $logger): void
     {
         $this->logger = $logger;
     }
 
-    public function log($message, $context = [])
+    public function log(string $message, array $context = []): void
     {
         $this->logger->debug($message, $context);
     }
