@@ -16,7 +16,7 @@ class DbMysqlFunctions extends DbBaseFunctions
         $this->deliTableRight = '`';
     }
 
-    public function concat($str1, $str2 = null)
+    public function concat(string $str1, ?string $str2 = null): string
     {
         return "concat(" . implode(', ', func_get_args()) . ")";
     }
@@ -28,12 +28,8 @@ class DbMysqlFunctions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
-    public function limit($sql, $start, $qty = null)
+    public function limit(string $sql, int $start, int $qty = 50): string
     {
-        if (is_null($qty)) {
-            $qty = 50;
-        }
-
         if (stripos($sql, ' LIMIT ') === false) {
             $sql = $sql . " LIMIT x, y";
         }
@@ -51,7 +47,7 @@ class DbMysqlFunctions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
-    public function top($sql, $qty)
+    public function top(string $sql, int $qty): string
     {
         return $this->limit($sql, 0, $qty);
     }
@@ -60,7 +56,7 @@ class DbMysqlFunctions extends DbBaseFunctions
      * Return if the database provider have a top or similar function
      * @return bool
      */
-    public function hasTop()
+    public function hasTop(): bool
     {
         return true;
     }
@@ -69,7 +65,7 @@ class DbMysqlFunctions extends DbBaseFunctions
      * Return if the database provider have a limit function
      * @return bool
      */
-    public function hasLimit()
+    public function hasLimit(): bool
     {
         return true;
     }
@@ -82,7 +78,7 @@ class DbMysqlFunctions extends DbBaseFunctions
      * @return string
      * @example $db->getDbFunctions()->SQLDate("d/m/Y H:i", "dtcriacao")
      */
-    public function sqlDate($format, $column = null)
+    public function sqlDate(string $format, ?string $column = null): string
     {
         if (is_null($column)) {
             $column = 'now()';
@@ -118,10 +114,10 @@ class DbMysqlFunctions extends DbBaseFunctions
      *
      * @param DbDriverInterface $dbdataset
      * @param string $sql
-     * @param array $param
-     * @return int
+     * @param array|null $param
+     * @return mixed
      */
-    public function executeAndGetInsertedId(DbDriverInterface $dbdataset, $sql, $param)
+    public function executeAndGetInsertedId(DbDriverInterface $dbdataset, string $sql, ?array $param = null): mixed
     {
         $returnedId = parent::executeAndGetInsertedId($dbdataset, $sql, $param);
         $iterator = $dbdataset->getIterator("select LAST_INSERT_ID() id");
@@ -133,12 +129,12 @@ class DbMysqlFunctions extends DbBaseFunctions
         return $returnedId;
     }
 
-    public function hasForUpdate()
+    public function hasForUpdate(): bool
     {
         return true;
     }
 
-    public function getTableMetadata(DbDriverInterface $dbdataset, $tableName)
+    public function getTableMetadata(DbDriverInterface $dbdataset, string $tableName): array
     {
         $sql = "EXPLAIN " . $this->deliTableLeft . $tableName . $this->deliTableRight;
         return $this->getTableMetadataFromSql($dbdataset, $sql);
@@ -160,19 +156,14 @@ class DbMysqlFunctions extends DbBaseFunctions
         return $return;
     }
 
-    public function getIsolationLevelCommand($isolationLevel)
+    public function getIsolationLevelCommand(?IsolationLevelEnum $isolationLevel = null): string
     {
-        switch ($isolationLevel) {
-            case IsolationLevelEnum::READ_UNCOMMITTED:
-                return "SET SESSION TRANSACTION ISOLATION LEVEL READ UNCOMMITTED";
-            case IsolationLevelEnum::READ_COMMITTED:
-                return "SET SESSION TRANSACTION ISOLATION LEVEL READ COMMITTED";
-            case IsolationLevelEnum::REPEATABLE_READ:
-                return "SET SESSION TRANSACTION ISOLATION LEVEL REPEATABLE READ";
-            case IsolationLevelEnum::SERIALIZABLE:
-                return "SET SESSION TRANSACTION ISOLATION LEVEL SERIALIZABLE";
-            default:
-                return "";
-        }
+        return match ($isolationLevel) {
+            IsolationLevelEnum::READ_UNCOMMITTED => "SET SESSION TRANSACTION ISOLATION LEVEL READ UNCOMMITTED",
+            IsolationLevelEnum::READ_COMMITTED => "SET SESSION TRANSACTION ISOLATION LEVEL READ COMMITTED",
+            IsolationLevelEnum::REPEATABLE_READ => "SET SESSION TRANSACTION ISOLATION LEVEL REPEATABLE READ",
+            IsolationLevelEnum::SERIALIZABLE => "SET SESSION TRANSACTION ISOLATION LEVEL SERIALIZABLE",
+            default => "",
+        };
     }
 }
