@@ -16,7 +16,7 @@ class DbOci8Functions extends DbBaseFunctions
         $this->deliTableRight = '"';
     }
 
-    public function concat($str1, $str2 = null)
+    public function concat(string $str1, ?string $str2 = null): string
     {
         return implode(' || ', func_get_args());
     }
@@ -28,12 +28,8 @@ class DbOci8Functions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
-    public function limit($sql, $start, $qty = null)
+    public function limit(string $sql, int $start, int $qty = 50): string
     {
-        if (is_null($qty)) {
-            $qty = 50;
-        }
-
         if (stripos($sql, ' OFFSET ') === false && stripos($sql, ' FETCH NEXT ') === false) {
             $sql = $sql . " OFFSET x ROWS FETCH NEXT y ROWS ONLY";
         }
@@ -51,7 +47,7 @@ class DbOci8Functions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
-    public function top($sql, $qty)
+    public function top(string $sql, int $qty): string
     {
         return $this->limit($sql, 0, $qty);
     }
@@ -60,7 +56,7 @@ class DbOci8Functions extends DbBaseFunctions
      * Return if the database provider have a top or similar function
      * @return bool
      */
-    public function hasTop()
+    public function hasTop(): bool
     {
         return true;
     }
@@ -69,7 +65,7 @@ class DbOci8Functions extends DbBaseFunctions
      * Return if the database provider have a limit function
      * @return bool
      */
-    public function hasLimit()
+    public function hasLimit(): bool
     {
         return true;
     }
@@ -82,7 +78,7 @@ class DbOci8Functions extends DbBaseFunctions
      * @return string
      * @example $db->getDbFunctions()->SQLDate("d/m/Y H:i", "dtcriacao")
      */
-    public function sqlDate($format, $column = null)
+    public function sqlDate(string $format, ?string $column = null): string
     {
         if (is_null($column)) {
             $column = 'current_timestamp';
@@ -115,10 +111,10 @@ class DbOci8Functions extends DbBaseFunctions
     /**
      * @param DbDriverInterface $dbdataset
      * @param string $sql
-     * @param array $param
-     * @return int
+     * @param array|null $param
+     * @return mixed
      */
-    public function executeAndGetInsertedId(DbDriverInterface $dbdataset, $sql, $param)
+    public function executeAndGetInsertedId(DbDriverInterface $dbdataset, string $sql, ?array $param = null): mixed
     {
         preg_match('/INSERT INTO ([a-zA-Z0-9_]+)/i', $sql, $matches);
         $tableName = $matches[1] ?? null;
@@ -160,12 +156,12 @@ class DbOci8Functions extends DbBaseFunctions
 
     }
 
-    public function hasForUpdate()
+    public function hasForUpdate(): bool
     {
         return true;
     }
 
-    public function getTableMetadata(DbDriverInterface $dbdataset, $tableName)
+    public function getTableMetadata(DbDriverInterface $dbdataset, string $tableName): array
     {
         $tableName = strtoupper($tableName);
         $sql = "SELECT
@@ -200,18 +196,13 @@ class DbOci8Functions extends DbBaseFunctions
         return $return;
     }
 
-    public function getIsolationLevelCommand($isolationLevel)
+    public function getIsolationLevelCommand(?IsolationLevelEnum $isolationLevel = null): string
     {
-        switch ($isolationLevel) {
-            case IsolationLevelEnum::READ_UNCOMMITTED:
-                return "SET TRANSACTION READ WRITE";
-            case IsolationLevelEnum::READ_COMMITTED:
-            case IsolationLevelEnum::REPEATABLE_READ:
-                return "SET TRANSACTION ISOLATION LEVEL READ COMMITTED";
-            case IsolationLevelEnum::SERIALIZABLE:
-                return "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE";
-            default:
-                return "";
-        }
+        return match ($isolationLevel) {
+            IsolationLevelEnum::READ_UNCOMMITTED => "SET TRANSACTION READ WRITE",
+            IsolationLevelEnum::READ_COMMITTED, IsolationLevelEnum::REPEATABLE_READ => "SET TRANSACTION ISOLATION LEVEL READ COMMITTED",
+            IsolationLevelEnum::SERIALIZABLE => "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE",
+            default => "",
+        };
     }
 }
