@@ -15,16 +15,16 @@ class SqlBind
      * Each provider have your own model for pass parameter.
      * This method define how each provider name define the parameters
      *
+     * The default is ":_"
+     *
+     * The symbol "_" will be replaced by the parameter name
+     *
      * @param Uri $connData
      * @return string
      */
-    public static function getParamModel(Uri $connData)
+    public static function getParamModel(Uri $connData): string
     {
-        if ($connData->getQueryPart("parammodel") != "") {
-            return $connData->getQueryPart("parammodel");
-        }
-
-        return ":_";
+        return $connData->getQueryPart("parammodel") ?? ":_";
     }
 
     /**
@@ -33,16 +33,16 @@ class SqlBind
      *
      * @param Uri $connData
      * @param string $sql
-     * @param array $params
+     * @param array|null $params
      * @return array An array with the adjusted SQL and PARAMs
      */
-    public static function parseSQL(Uri $connData, $sql, $params = null)
+    public static function parseSQL(Uri $connData, string $sql, array $params = null): array
     {
         $paramSubstName = SqlBind::getParamModel($connData);
 
         $sqlAlter = preg_replace("~'.*?((\\\\'|'').*?)*'~", "", $sql);
         preg_match_all(
-            "/(?<deliStart>\\[\\[|:)(?<param>[\\w\\d]+)(?<deliEnd>\\]\\]|[^\\d\\w]|$)/",
+            "/:(?<param>[_\\w\\d]+)\b/",
             $sqlAlter,
             $matches
         );
@@ -58,30 +58,26 @@ class SqlBind
                 // Remove NON DEFINED parameters
                 $sql = preg_replace(
                     [
-                        "/\\[\\[$paramName\\]\\]/",
-                        "/:$paramName([^\\d\\w]|$)/"
+                        "/:$paramName\b/"
                     ],
                     [
-                        "null",
-                        "null$2"
+                        "null"
                     ],
                     $sql
                 );
                 continue;
             }
 
-            $usedParams[$paramName] = isset($params[$paramName]) ? $params[$paramName] : null;
+            $usedParams[$paramName] = $params[$paramName] ?? null;
             $dbArg = str_replace("_", SqlBind::keyAdj($paramName), $paramSubstName);
 
             $count = 0;
             $sql = preg_replace(
                 [
-                    "/\\[\\[$paramName\\]\\]/",
-                    "/:$paramName([^\\w\\d]|$)/",
+                    "/:$paramName\b/",
                 ],
                 [
-                    $dbArg . '',
-                    $dbArg . '$1',
+                    $dbArg,
                 ],
                 $sql,
                 -1,
@@ -92,7 +88,7 @@ class SqlBind
         return [$sql, $usedParams];
     }
 
-    public static function keyAdj($key)
+    public static function keyAdj(string $key): string
     {
         return str_replace(".", "_", $key);
     }
