@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\AnyDataset\Sample;
+namespace Test;
 
 use ByJG\Serializer\BaseModel;
 
@@ -10,15 +10,16 @@ use ByJG\Serializer\BaseModel;
 class LiteralSample extends BaseModel
 {
 
-    protected $value = "";
+    protected string $value = "";
 
     /**
      * LiteralSample constructor.
      *
-     * @param string $value
+     * @param string|int $value
      */
-    public function __construct($value)
+    public function __construct(string|int $value)
     {
+        parent::__construct();
         $this->value = $value;
     }
 
