@@ -4,10 +4,10 @@ namespace ByJG\AnyDataset\Db\Interfaces;
 
 interface DbCacheInterface
 {
-    public function getMaxStmtCache();
+    public function getMaxStmtCache(): int;
 
-    public function getCountStmtCache();
+    public function getCountStmtCache(): int;
 
-    public function setMaxStmtCache($maxStmtCache);
+    public function setMaxStmtCache(int $maxStmtCache): void;
 
 }
