@@ -8,7 +8,7 @@ use ByJG\AnyDataset\Db\Helpers\SqlHelper;
 
 class IteratorFilterSqlFormatter extends IteratorFilterFormatter
 {
-    public function format($filters, $tableName = null, &$params = [], $returnFields = "*")
+    public function format(array $filters, string $tableName = null, array &$params = [], string $returnFields = "*"): string
     {
         $params = array();
 
@@ -28,7 +28,7 @@ class IteratorFilterSqlFormatter extends IteratorFilterFormatter
         );
     }
 
-    public function getRelation($name, $relation, $value, &$param)
+    public function getRelation(string $name, Relation $relation, mixed $value, array &$param): string
     {
         $paramName = $name;
         $counter = 0;
@@ -38,7 +38,7 @@ class IteratorFilterSqlFormatter extends IteratorFilterFormatter
 
         $paramStr = function (&$param, $paramName, $value) {
             $param[$paramName] = trim($value);
-            $result = "[[$paramName]]";
+            $result = ":$paramName";
             if (is_object($value)) {
                 unset($param[$paramName]);
                 $result = $value->__toString();
@@ -46,42 +46,48 @@ class IteratorFilterSqlFormatter extends IteratorFilterFormatter
             return $result;
         };
 
-        $data = [
+        $data = match ($relation) {
             Relation::EQUAL => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 return " $name = " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::GREATER_THAN => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 return " $name > " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::LESS_THAN => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 return " $name < " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::GREATER_OR_EQUAL_THAN => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 return " $name >= " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::LESS_OR_EQUAL_THAN => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 return " $name <= " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::NOT_EQUAL => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 return " $name <> " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::STARTS_WITH => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 $value .= "%";
                 return " $name  like  " . $paramStr($param, $paramName, $value) . ' ';
             },
-
             Relation::CONTAINS => function (&$param, $name, $paramName, $value) use ($paramStr) {
                 $value = "%" . $value . "%";
                 return " $name  like  " . $paramStr($param, $paramName, $value) . ' ';
-            }
-        ];
+            },
+            Relation::IN => function (&$param, $name, $paramName, $value) {
+                $placeholders = implode(', ', array_map(fn($v, $i) => ":$paramName$i", $value, array_keys($value)));
+                foreach ($value as $i => $v) {
+                    $param["$paramName$i"] = $v;
+                }
+                return " $name IN ($placeholders) ";
+            },
+            Relation::NOT_IN => function (&$param, $name, $paramName, $value) {
+                $placeholders = implode(', ', array_map(fn($v, $i) => ":$paramName$i", $value, array_keys($value)));
+                foreach ($value as $i => $v) {
+                    $param["$paramName$i"] = $v;
+                }
+                return " $name NOT IN ($placeholders) ";
+            },
+        };
 
-        return $data[$relation]($param, $name, $paramName, $value);
-    }
+        return $data($param, $name, $paramName, $value);    }
 }
