@@ -1,14 +1,11 @@
 <?php
 
-namespace Tests\AnyDataset\Dataset;
+namespace Test;
 
+use ByJG\AnyDataset\Core\Enum\Relation;
 use ByJG\AnyDataset\Core\IteratorFilter;
 use ByJG\AnyDataset\Db\IteratorFilterSqlFormatter;
-use ByJG\AnyDataset\Core\Enum\Relation;
 use PHPUnit\Framework\TestCase;
-use Tests\AnyDataset\Sample\LiteralSample;
-
-require_once 'LiteralSample.php';
 
 class IteratorFilterTest extends TestCase
 {
@@ -29,7 +26,7 @@ class IteratorFilterTest extends TestCase
 
     public function testGetSql()
     {
-        $params = null;
+        $params = [];
         $returnFields = '*';
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
@@ -40,7 +37,7 @@ class IteratorFilterTest extends TestCase
         $this->assertEquals([], $params);
         $this->assertEquals('select * from tablename ', $sql);
 
-        $this->object->addRelation('field', Relation::EQUAL, 'test');
+        $this->object->and('field', Relation::EQUAL, 'test');
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
             'tablename',
@@ -48,9 +45,9 @@ class IteratorFilterTest extends TestCase
             $returnFields
         );
         $this->assertEquals(['field' => 'test'], $params);
-        $this->assertEquals('select * from tablename  where  field = [[field]]  ', $sql);
+        $this->assertEquals('select * from tablename  where  field = :field  ', $sql);
 
-        $this->object->addRelation('field2', Relation::GREATER_OR_EQUAL_THAN, 'test2');
+        $this->object->and('field2', Relation::GREATER_OR_EQUAL_THAN, 'test2');
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
             'tablename',
@@ -58,9 +55,9 @@ class IteratorFilterTest extends TestCase
             $returnFields
         );
         $this->assertEquals(['field' => 'test', 'field2' => 'test2'], $params);
-        $this->assertEquals('select * from tablename  where  field = [[field]]  and  field2 >= [[field2]]  ', $sql);
+        $this->assertEquals('select * from tablename  where  field = :field  and  field2 >= :field2  ', $sql);
 
-        $this->object->addRelation('field3', Relation::CONTAINS, 'test3');
+        $this->object->and('field3', Relation::CONTAINS, 'test3');
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
             'tablename',
@@ -68,14 +65,14 @@ class IteratorFilterTest extends TestCase
             $returnFields
         );
         $this->assertEquals(['field' => 'test', 'field2' => 'test2', 'field3' => '%test3%'], $params);
-        $this->assertEquals('select * from tablename  where  field = [[field]]  and  field2 >= [[field2]]  and  field3  like  [[field3]]  ', $sql);
+        $this->assertEquals('select * from tablename  where  field = :field  and  field2 >= :field2  and  field3  like  :field3  ', $sql);
     }
 
     public function testSqlLiteral()
     {
         $literalObject = new LiteralSample(10);
 
-        $params = null;
+        $params = [];
         $returnFields = '*';
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
@@ -86,7 +83,7 @@ class IteratorFilterTest extends TestCase
         $this->assertEquals([], $params);
         $this->assertEquals('select * from tablename ', $sql);
 
-        $this->object->addRelation('field', Relation::GREATER_THAN, $literalObject);
+        $this->object->and('field', Relation::GREATER_THAN, $literalObject);
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
             'tablename',
@@ -96,7 +93,7 @@ class IteratorFilterTest extends TestCase
         $this->assertEquals([], $params);
         $this->assertEquals('select * from tablename  where  field > cast(\'10\' as integer)  ', $sql);
 
-        $this->object->addRelation('field2', Relation::LESS_THAN, 5);
+        $this->object->and('field2', Relation::LESS_THAN, 5);
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
             'tablename',
@@ -104,15 +101,45 @@ class IteratorFilterTest extends TestCase
             $returnFields
         );
         $this->assertEquals(['field2' => 5], $params);
-        $this->assertEquals('select * from tablename  where  field > cast(\'10\' as integer)  and  field2 < [[field2]]  ', $sql);
+        $this->assertEquals('select * from tablename  where  field > cast(\'10\' as integer)  and  field2 < :field2  ', $sql);
+    }
+
+    public function testRelationIn()
+    {
+        $this->object->and('field', Relation::IN, ['value1', 'value2']);
+        $params = [];
+        $returnFields = '*';
+        $sql = $this->object->format(
+            new IteratorFilterSqlFormatter(),
+            'tablename',
+            $params,
+            $returnFields
+        );
+        $this->assertEquals(['field0' => 'value1', 'field1' => 'value2'], $params);
+        $this->assertEquals('select * from tablename  where  field IN (:field0, :field1)  ', $sql);
+    }
+
+    public function testRelationNotIn()
+    {
+        $this->object->and('field', Relation::NOT_IN, ['value1', 'value2']);
+        $params = [];
+        $returnFields = '*';
+        $sql = $this->object->format(
+            new IteratorFilterSqlFormatter(),
+            'tablename',
+            $params,
+            $returnFields
+        );
+        $this->assertEquals(['field0' => 'value1', 'field1' => 'value2'], $params);
+        $this->assertEquals('select * from tablename  where  field NOT IN (:field0, :field1)  ', $sql);
     }
 
     public function testAddRelationOr()
     {
-        $this->object->addRelation('field', Relation::EQUAL, 'test');
-        $this->object->addRelationOr('field2', Relation::EQUAL, 'test2');
+        $this->object->and('field', Relation::EQUAL, 'test');
+        $this->object->or('field2', Relation::EQUAL, 'test2');
 
-        $params = null;
+        $params = [];
         $returnFields = '*';
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
@@ -121,18 +148,18 @@ class IteratorFilterTest extends TestCase
             $returnFields
         );
         $this->assertEquals(['field' => 'test', 'field2' => 'test2'], $params);
-        $this->assertEquals('select * from tablename  where  field = [[field]]  or  field2 = [[field2]]  ', $sql);
+        $this->assertEquals('select * from tablename  where  field = :field  or  field2 = :field2  ', $sql);
     }
 
     public function testGroup()
     {
         $this->object->startGroup();
-        $this->object->addRelation('field', Relation::EQUAL, 'test');
-        $this->object->addRelation('field2', Relation::EQUAL, 'test2');
+        $this->object->and('field', Relation::EQUAL, 'test');
+        $this->object->and('field2', Relation::EQUAL, 'test2');
         $this->object->endGroup();
-        $this->object->addRelationOr('field3', Relation::EQUAL, 'test3');
+        $this->object->or('field3', Relation::EQUAL, 'test3');
 
-        $params = null;
+        $params = [];
         $returnFields = '*';
         $sql = $this->object->format(
             new IteratorFilterSqlFormatter(),
@@ -142,7 +169,7 @@ class IteratorFilterTest extends TestCase
         );
         $this->assertEquals(['field' => 'test', 'field2' => 'test2', 'field3' => 'test3'], $params);
         $this->assertEquals(
-            'select * from tablename  where  (  field = [[field]]  and  field2 = [[field2]] ) or  field3 = [[field3]]  ',
+            'select * from tablename  where  (  field = :field  and  field2 = :field2 ) or  field3 = :field3  ',
             $sql
         );
     }
