@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\AnyDataset\Store;
+namespace Test;
 
 use ByJG\AnyDataset\Db\Factory;
 use ByJG\AnyDataset\Db\Helpers\DbSqliteFunctions;
@@ -17,7 +17,7 @@ class PdoSqliteTest extends TestCase
 
     public function setUp(): void
     {
-        $this->dbDriver = Factory::getDbRelationalInstance('sqlite:///tmp/test.db');
+        $this->dbDriver = Factory::getDbInstance('sqlite:///tmp/test.db');
 
         $this->dbDriver->execute(
             'create table users (
@@ -47,6 +47,7 @@ class PdoSqliteTest extends TestCase
         unlink('/tmp/test.db');
     }
 
+    /** @psalm-suppress InvalidArrayOffset */
     public function testGetIterator()
     {
         $iterator = $this->dbDriver->getIterator('select * from info');
@@ -79,6 +80,7 @@ class PdoSqliteTest extends TestCase
         }
     }
 
+    /** @psalm-suppress InvalidArrayOffset */
     public function testGetIteratorFilter()
     {
         $iterator = $this->dbDriver->getIterator('select * from info where iduser = :id', ['id' => 1]);
@@ -156,7 +158,7 @@ class PdoSqliteTest extends TestCase
     public function testExecute()
     {
         $this->dbDriver->execute("insert into users (name, createdate) values ('Another', '2017-05-11')");
-        $iterator = $this->dbDriver->getIterator('select * from users where name = [[name]]', ['name' => 'Another']);
+        $iterator = $this->dbDriver->getIterator('select * from users where name = :name', ['name' => 'Another']);
 
         $this->assertEquals(
             [
@@ -171,7 +173,7 @@ class PdoSqliteTest extends TestCase
         $newId = $this->dbDriver->executeAndGetId("insert into users (name, createdate) values ('Another', '2017-05-11')");
 
         $this->assertEquals(4, $newId);
-        $iterator = $this->dbDriver->getIterator('select * from users where name = [[name]]', ['name' => 'Another']);
+        $iterator = $this->dbDriver->getIterator('select * from users where name = :name', ['name' => 'Another']);
 
         $this->assertEquals(
             [
@@ -194,7 +196,7 @@ class PdoSqliteTest extends TestCase
         $this->assertEquals(4, $newId);
         $this->dbDriver->commitTransaction();
 
-        $iterator = $this->dbDriver->getIterator('select * from users where name = [[name]]', ['name' => 'Another']);
+        $iterator = $this->dbDriver->getIterator('select * from users where name = :name', ['name' => 'Another']);
 
         $this->assertEquals(
             [
@@ -211,7 +213,7 @@ class PdoSqliteTest extends TestCase
         $this->assertEquals(4, $newId);
         $this->dbDriver->rollbackTransaction();
 
-        $iterator = $this->dbDriver->getIterator('select * from users where name = [[name]]', ['name' => 'Another']);
+        $iterator = $this->dbDriver->getIterator('select * from users where name = :name', ['name' => 'Another']);
         $this->assertFalse($iterator->hasNext());
     }
 
@@ -225,17 +227,17 @@ class PdoSqliteTest extends TestCase
         $this->dbDriver->rollbackTransaction();
 
         // Context 2
-        $context2 = Factory::getDbRelationalInstance('sqlite:///tmp/test.db');
+        $context2 = Factory::getDbInstance('sqlite:///tmp/test.db');
         $context2->beginTransaction();
         $newId = $context2->executeAndGetId("insert into users (name, createdate) values ('Another2', '2017-04-11')");
         $this->assertEquals(4, $newId);
         $context2->commitTransaction();
 
         // Check values
-        $iterator = $this->dbDriver->getIterator('select * from users where name = [[name]]', ['name' => 'Another']);
+        $iterator = $this->dbDriver->getIterator('select * from users where name = :name', ['name' => 'Another']);
         $this->assertFalse($iterator->hasNext());
 
-        $iterator = $this->dbDriver->getIterator('select * from users where name = [[name]]', ['name' => 'Another2']);
+        $iterator = $this->dbDriver->getIterator('select * from users where name = :name', ['name' => 'Another2']);
         $this->assertEquals(
             [
                 ['id' => 4, 'name' => 'Another2', 'createdate' => '2017-04-11'],
@@ -266,7 +268,7 @@ class PdoSqliteTest extends TestCase
 
     public function testisSupportMultRowset()
     {
-        $this->assertFalse($this->dbDriver->isSupportMultRowset());
+        $this->assertFalse($this->dbDriver->isSupportMultiRowset());
     }
 
     public function testCachedResults()
@@ -282,8 +284,8 @@ class PdoSqliteTest extends TestCase
             $iterator->toArray()
         );
 
-        // Remove it from DB (Still in cache) - Execute dont use cache
-        $this->dbDriver->execute("delete from users where name = [[name]]", ['name' => 'Another2'], $cache, 60);
+        // Remove it from DB (Still in cache) - Execute don't use cache
+        $this->dbDriver->execute("delete from users where name = :name", ['name' => 'Another2']);
 
         // Try get from cache
         $iterator = $this->dbDriver->getIterator('select * from info where id = :id', ['id' => 1], $cache, 60);
